feat!: rename the realtime_client package to supabase_realtime
The package name says nothing about Supabase and does not match how the
rest of the packages are named. It is published as supabase_realtime
from v3 onwards, with the library entrypoint renamed to match.

No public types are renamed, so this is a package rename only. The
X-Client-Info header still reports realtime-dart for telemetry
continuity, matching the decision made for supabase_auth.
