+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -25,7 +25,7 @@ body:
         - supabase
         - supabase_auth
         - postgrest
-        - realtime_client
+        - supabase_realtime
         - storage_client
         - supabase_functions
         - yet_another_json_isolate
@@ -58,10 +58,10 @@ body:
         Output of `flutter doctor -v`, plus the versions of the affected packages.
 
         On Linux/macOS:
-        `dart pub deps | grep -E "supabase|postgrest|storage_client|realtime_client"`
+        `dart pub deps | grep -E "supabase|postgrest|storage_client"`
 
         On Windows:
-        `dart pub deps | findstr "supabase postgrest storage_client realtime_client"`
+        `dart pub deps | findstr "supabase postgrest storage_client"`
       render: shell
       placeholder: Flutter, Dart and package versions
     validations:
