+++ MIGRATION.md
@@ -4,9 +4,10 @@ This document describes the breaking changes you need to be aware of when upgrad
 versions of the Supabase Flutter SDK, together with the steps required to migrate your code.
 
 All packages in this repository are released together for a major version, so a single section
-covers `supabase_flutter`, `supabase`, `supabase_auth`, `postgrest`, `realtime_client`, `storage_client`
-and `supabase_functions`. Every symbol mentioned here is re-exported from `supabase_flutter`, so the
-snippets apply whether you depend on the individual package or on the Flutter one.
+covers `supabase_flutter`, `supabase`, `supabase_auth`, `postgrest`, `supabase_realtime`,
+`storage_client` and `supabase_functions`. Every symbol mentioned here is re-exported from
+`supabase_flutter`, so the snippets apply whether you depend on the individual package or on the
+Flutter one.
 
 ## Migrating from v2 to v3
 
@@ -147,6 +148,38 @@ No types are renamed, `FunctionsClient` and everything around it keep their name
 the wire format changes either: the `X-Client-Info` header still identifies this client as
 `functions-dart`.
 
+### The `realtime_client` package is now `supabase_realtime`
+
+`realtime_client` says nothing about Supabase and does not match how the rest of the packages are
+named. The package is published as `supabase_realtime` from v3 onwards, and the `realtime_client`
+package is discontinued on pub.dev when v3 ships.
+
+If you depend on `supabase_flutter` or `supabase` you do not need to change anything, both pull in
+`supabase_realtime` for you and re-export it.
+
+If you depend on the realtime client directly, rename the dependency and the import:
+
+```yaml
+# Before
+dependencies:
+  realtime_client: ^2.13.0
+
+# After
+dependencies:
+  supabase_realtime: ^3.0.0
+```
+
+```dart
+// Before
+import 'package:realtime_client/realtime_client.dart';
+
+// After
+import 'package:supabase_realtime/supabase_realtime.dart';
+```
+
+The rename does not touch any type names. `RealtimeClient`, `RealtimeChannel` and the rest keep
+their names, and the `X-Client-Info` header still identifies this client as `realtime-dart`.
+
 ### `RealtimeClient.connectionState` is now typed
 
 `RealtimeClient` used to expose the socket state twice: a typed `connState` field and a stringly
@@ -252,8 +285,8 @@ ones below are the renames you can actually hit.
 
 | Before | After | Package |
 | --- | --- | --- |
-| `SocketStates` | `SocketState` | `realtime_client` |
-| `PostgresTypes` | `PostgresType` | `realtime_client` |
+| `SocketStates` | `SocketState` | `supabase_realtime` |
+| `PostgresTypes` | `PostgresType` | `supabase_realtime` |
 | `AuthenticatorAssuranceLevels` | `AuthenticatorAssuranceLevel` | `supabase_auth` |
 | `LoadTableSnapshots` | `TableSnapshotScope` | `storage_client` |
 
@@ -439,8 +472,8 @@ already inert: unused types, options the client ignored, or values the server ne
 | `User.confirmedAt` | `User.emailConfirmedAt` | `supabase_auth` |
 | `ReturningOption` | none, it was unused | `postgrest` |
 | `PostgrestClient.auth()` | `PostgrestClient.setAuth()` | `postgrest` |
-| `RealtimeClient.longpollerTimeout` | none, there is no longpoll transport | `realtime_client` |
-| `ChannelResponse.rateLimited` | none, it was never returned | `realtime_client` |
+| `RealtimeClient.longpollerTimeout` | none, there is no longpoll transport | `supabase_realtime` |
+| `ChannelResponse.rateLimited` | none, it was never returned | `supabase_realtime` |
 | `FileObject.lastAccessedAt`, `FileObjectV2.lastAccessedAt` | none, the server does not populate it | `storage_client` |
 | `AuthUser` | `User` | `supabase` |
 | `RealtimeClientOptions.eventsPerSecond` | none, it was already ignored | `supabase` |
