--- packages/realtime_client/lib/src/version.dart
@@ -1,1 +1,0 @@
-const version = '2.13.0';
