+++ packages/supabase/lib/src/realtime_client_options.dart
@@ -1,4 +1,4 @@
-import 'package:realtime_client/realtime_client.dart';
+import 'package:supabase_realtime/supabase_realtime.dart';
 
 /// {@template realtime_client_options}
 /// Options to pass to the RealtimeClient.
