+++ packages/supabase/pubspec.yaml
@@ -21,11 +21,11 @@ dependencies:
   http: ^1.6.0
   meta: ^1.7.0
   postgrest: 2.9.1
-  realtime_client: 2.13.0
   storage_client: 2.8.0
   supabase_auth: 3.0.0-dev.1
   supabase_common: 0.1.2
   supabase_functions: 3.0.0-dev.1
+  supabase_realtime: 3.0.0-dev.1
   yet_another_json_isolate: 2.1.1
   logging: ^1.3.0
