+++ packages/supabase_realtime/lib/src/message.dart
@@ -1,5 +1,5 @@
-import 'package:realtime_client/realtime_client.dart';
-import 'package:realtime_client/src/constants.dart';
+import 'package:supabase_realtime/supabase_realtime.dart';
+import 'package:supabase_realtime/src/constants.dart';
 import 'package:meta/meta.dart';
 
 @internal
