+++ packages/supabase_realtime/lib/src/push.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
-import 'package:realtime_client/realtime_client.dart';
-import 'package:realtime_client/src/constants.dart';
-import 'package:realtime_client/src/message.dart';
-import 'package:realtime_client/src/types.dart';
+import 'package:supabase_realtime/supabase_realtime.dart';
+import 'package:supabase_realtime/src/constants.dart';
+import 'package:supabase_realtime/src/message.dart';
+import 'package:supabase_realtime/src/types.dart';
 import 'package:meta/meta.dart';
 
 @internal
