+++ packages/supabase_realtime/lib/src/realtime_channel.dart
@@ -5,12 +5,12 @@ import 'dart:typed_data';
 import 'package:collection/collection.dart';
 import 'package:http/http.dart';
 import 'package:meta/meta.dart';
-import 'package:realtime_client/realtime_client.dart';
-import 'package:realtime_client/src/constants.dart';
-import 'package:realtime_client/src/push.dart';
-import 'package:realtime_client/src/retry_timer.dart';
-import 'package:realtime_client/src/transformers.dart';
-import 'package:realtime_client/src/types.dart';
+import 'package:supabase_realtime/supabase_realtime.dart';
+import 'package:supabase_realtime/src/constants.dart';
+import 'package:supabase_realtime/src/push.dart';
+import 'package:supabase_realtime/src/retry_timer.dart';
+import 'package:supabase_realtime/src/transformers.dart';
+import 'package:supabase_realtime/src/types.dart';
 
 class RealtimeChannel {
   final Map<String, List<Binding>> _bindings = {};
