+++ packages/supabase_realtime/lib/src/realtime_presence.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:realtime_client/realtime_client.dart';
-import 'package:realtime_client/src/types.dart';
+import 'package:supabase_realtime/supabase_realtime.dart';
+import 'package:supabase_realtime/src/types.dart';
 
 /// A single shared state between users with Realtime Presence.
 class Presence {
