+++ packages/supabase_realtime/lib/src/serializer.dart
