+++ packages/supabase_realtime/lib/src/types.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
-import 'package:realtime_client/realtime_client.dart';
+import 'package:supabase_realtime/supabase_realtime.dart';
 
 typedef BindingCallback = void Function(dynamic payload, [dynamic ref]);
