+++ packages/supabase_realtime/lib/src/version.dart
@@ -1,0 +1,1 @@
+const version = '3.0.0-dev.1';
