+++ packages/supabase_realtime/lib/src/websocket/websocket.dart
