+++ packages/supabase_realtime/lib/src/websocket/websocket_stub.dart
