+++ packages/supabase_realtime/lib/src/websocket/websocket_web.dart
