+++ packages/supabase_realtime/pubspec.yaml
@@ -1,8 +1,8 @@
-name: realtime_client
+name: supabase_realtime
 description: Listens to changes in a PostgreSQL Database and via websockets. This is for usage with Supabase Realtime server.
-version: 2.13.0
+version: 3.0.0-dev.1
 homepage: 'https://supabase.com'
-repository: 'https://github.com/supabase/supabase-flutter/tree/main/packages/realtime_client'
+repository: 'https://github.com/supabase/supabase-flutter/tree/main/packages/supabase_realtime'
 issue_tracker: 'https://github.com/supabase/supabase-flutter/issues'
 documentation: 'https://supabase.com/docs/reference/dart/subscribe'
 topics:
