+++ packages/supabase_realtime/test/socket_test_stubs.dart
@@ -1,6 +1,6 @@
 import 'package:mocktail/mocktail.dart';
-import 'package:realtime_client/realtime_client.dart';
-import 'package:realtime_client/src/push.dart';
+import 'package:supabase_realtime/supabase_realtime.dart';
+import 'package:supabase_realtime/src/push.dart';
 import 'package:web_socket_channel/io.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
