+++ packages/supabase_realtime/test/types_test.dart
@@ -1,4 +1,4 @@
-import 'package:realtime_client/realtime_client.dart';
+import 'package:supabase_realtime/supabase_realtime.dart';
 import 'package:test/test.dart';
 
 Map<String, dynamic> payloadWith(Object? commitTimestamp) => {
